When asked for a file that doesn't exist, check it.
Currently, AzureFS cached container contents, but
never really invalidates it. When a new file is put
to Azure by some other host (not us), we won't see it.

With this patch, when asked for a file we don't know
about, we query Azure whenever this file exists.
There is a negative reply 30-seconds cache (currently,
hardcoded), so we don't hammer API with requests
too frequenty.

Also, this patch adds a hack to ignore 0-sized blobs.
This is a temporary measure/hack to prevent us from
prematurely caching (forever!) pending upload entries.
It should be removed, as proper a cache invalidation
is implemented.